refactor: - Introduced `validate_addon.bat` for addon validation. - Updated `blender_manifest.toml` with new version and website. - Created `config.py` for default settings and prompts. - Implemented `DeepSeekProperties` in `properties.py` for user configuration. - Added `DEEPSEEK_OT_FixErrors` operator in `fix_errors.py` to fix Python errors using AI. - Added `DEEPSEEK_OT_AutoComplete` operator in `autocomplete.py` for generating code suggestions.
