+++ source/validate_addon.bat
@@ -1,0 +1,1 @@
+"C:\Program Files\Blender Foundation\Blender 4.2\blender.exe" -c extension validate
